requirements.txt
PyPDF2>=3.0.0
Pillow>=10.0.0
tabula-py>=2.8.0
pygame>=2.5.0
